Use `scripts.h` for platform specific commands -
This should be easier to maintain all specifics
of different platforms and platform options. We
now alos introduc support for `ip` command, which
if a modern replacement to `ifconfig`/`route` on
Linux (and possibly other platfroms).

At the moment it defines macros for each of the
scripts, but this could be done a differently if
it gets tedious. It also inlines the functions
and renames `cleanup()` to `tundown()` as well
as `ifconf()` to `tunconf()`.