Use external scripts -
embedding shell code at compile time make it very unflexible,
we should supply to scripts as argumets, so the sequence of
commands to configure and take down the interface can be tailored
towards the target system. There two parameters which the script
has to know (`SLIP_CONFIG_TUNDEV` and `SLIP_CONFIG_IPADDR`), these
are passed via the environment.